refactor(desktop): extract useMailActions from App.tsx (F5.3)
Move the message-mutation cluster into useMailActions: single-message
archive/trash/read/unread, the bulk equivalents (bulkActionIds/bulkAction),
list add/remove/relabel, selection toggle, and the cursor-advance logic
(advanceAfterBulk/clearReaderIfRemoved) that keeps the reader in sync.

Behavior-preserving verbatim move. Bulk state (bulkMode/selected/
bulkProgress) stays in App and is passed in, so its many other readers
(render, commands, AI bulk-prompt) are untouched.

- useMailActions.ts (335); App.tsx 3,937 -> 3,689
- new mailactions.spec.ts guards single + bulk archive (previously
  uncovered); unit 61 / e2e 36, tsc + build clean
